feat/implemented-dropdown-menu-and-changed-logout-from-profile-to-dropdown-menu
