+++ src/components/components/profile-page/ProfileDetails.vue
@@ -1,5 +1,4 @@
 <script>
-import router from '../../../router';
 // Import the stores.
 import { useUserDetailsStore } from '../../../stores/UserDetailsStore';
 import { useSessionDetailsStore } from '../../../stores/SessionDetailsStore.js';
@@ -27,25 +26,7 @@ export default {
                 : '';
         }
     },
-    methods: {
-        LogOut() {
-            // Remove background image from theme on logout.
-            document.getElementsByTagName('body')[0].style =
-                'background-image: none;';
-
-            this.sessionDetailsStore.isLoggedIn = false;
-
-            const requestOptions = {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' }
-            };
-            var url = '/logout';
-
-            fetch(url, requestOptions).then(function (response) {
-                router.push({ name: 'login' });
-            });
-        }
-    }
+    methods: {}
 };
 </script>
 
@@ -68,29 +49,9 @@ export default {
                         class="d-lg-none img-fluid"
                     />
                     <div class="d-md-none">
-                        <!-- Log in or out -->
+                        <!-- Log in -->
                         <a
-                            v-if="sessionDetailsStore.isLoggedIn == true"
-                            @click="LogOut()"
-                            class="btn red-btn"
-                            role="button"
-                        >
-                            Log out&nbsp;&nbsp;
-                            <!-- X icon -->
-                            <svg
-                                xmlns="http://www.w3.org/2000/svg"
-                                viewBox="0 0 512 512"
-                                width="22"
-                                height="22"
-                                fill="white"
-                            >
-                                <path
-                                    d="M377.9 105.9L500.7 228.7c7.2 7.2 11.3 17.1 11.3 27.3s-4.1 20.1-11.3 27.3L377.9 406.1c-6.4 6.4-15 9.9-24 9.9c-18.7 0-33.9-15.2-33.9-33.9l0-62.1-128 0c-17.7 0-32-14.3-32-32l0-64c0-17.7 14.3-32 32-32l128 0 0-62.1c0-18.7 15.2-33.9 33.9-33.9c9 0 17.6 3.6 24 9.9zM160 96L96 96c-17.7 0-32 14.3-32 32l0 256c0 17.7 14.3 32 32 32l64 0c17.7 0 32 14.3 32 32s-14.3 32-32 32l-64 0c-53 0-96-43-96-96L0 128C0 75 43 32 96 32l64 0c17.7 0 32 14.3 32 32s-14.3 32-32 32z"
-                                />
-                            </svg>
-                        </a>
-                        <a
-                            v-else
+                            v-if="sessionDetailsStore.isLoggedIn == false"
                             href="/login"
                             class="btn primary-btn"
                             role="button"
@@ -158,29 +119,9 @@ export default {
                             </svg>
                         </router-link>
                         <div>
-                            <!-- Log in or out -->
+                            <!-- Log in -->
                             <a
-                                v-if="sessionDetailsStore.isLoggedIn == true"
-                                @click="LogOut()"
-                                class="btn red-btn"
-                                role="button"
-                            >
-                                Log out&nbsp;&nbsp;
-                                <!-- X icon -->
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    viewBox="0 0 512 512"
-                                    width="22"
-                                    height="22"
-                                    fill="white"
-                                >
-                                    <path
-                                        d="M377.9 105.9L500.7 228.7c7.2 7.2 11.3 17.1 11.3 27.3s-4.1 20.1-11.3 27.3L377.9 406.1c-6.4 6.4-15 9.9-24 9.9c-18.7 0-33.9-15.2-33.9-33.9l0-62.1-128 0c-17.7 0-32-14.3-32-32l0-64c0-17.7 14.3-32 32-32l128 0 0-62.1c0-18.7 15.2-33.9 33.9-33.9c9 0 17.6 3.6 24 9.9zM160 96L96 96c-17.7 0-32 14.3-32 32l0 256c0 17.7 14.3 32 32 32l64 0c17.7 0 32 14.3 32 32s-14.3 32-32 32l-64 0c-53 0-96-43-96-96L0 128C0 75 43 32 96 32l64 0c17.7 0 32 14.3 32 32s-14.3 32-32 32z"
-                                    />
-                                </svg>
-                            </a>
-                            <a
-                                v-else
+                                v-if="sessionDetailsStore.isLoggedIn == false"
                                 href="/login"
                                 class="btn primary-btn"
                                 role="button"
