+++ src/components/pages/LearningTracksView.vue
@@ -254,7 +254,7 @@ export default {
 .legend-div {
     width: 100%;
     z-index: 2;
-    top: 70px;
+    top: 80px;
 }
 
 /* Modal filters */
