+++ src/components/pages/RadialTreeView.vue
@@ -114,7 +114,7 @@ export default {
                 this.markTutorialComplete();
             }
         },
-        restartTutorial(){
+        restartTutorial() {
             this.showTutorialTip1 = true;
             this.isTutorialComplete = false;
         },
@@ -172,28 +172,31 @@ export default {
                         :clearResults="clearResult"
                     />
                     <!-- Print Button -->
-                     <div>
+                    <div>
                         <button
-                        class="btn primary-btn me-2"
-                        @click="$refs.childComponent.printPDF()"
-                    >
-                        Print
-                    </button>
-                    <button class="btn primary-btn" @click="restartTutorial">
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            viewBox="0 0 192 512"
-                            width="20"
-                            height="20"
-                            fill="white"
+                            class="btn primary-btn me-2"
+                            @click="$refs.childComponent.printPDF()"
                         >
-                            <!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc. -->
-                            <path
-                                d="M48 80a48 48 0 1 1 96 0A48 48 0 1 1 48 80zM0 224c0-17.7 14.3-32 32-32l64 0c17.7 0 32 14.3 32 32l0 224 32 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 512c-17.7 0-32-14.3-32-32s14.3-32 32-32l32 0 0-192-32 0c-17.7 0-32-14.3-32-32z"
-                            />
-                        </svg>
-                    </button>
-                     </div>          
+                            Print
+                        </button>
+                        <button
+                            class="btn primary-btn"
+                            @click="restartTutorial"
+                        >
+                            <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                viewBox="0 0 192 512"
+                                width="20"
+                                height="20"
+                                fill="white"
+                            >
+                                <!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc. -->
+                                <path
+                                    d="M48 80a48 48 0 1 1 96 0A48 48 0 1 1 48 80zM0 224c0-17.7 14.3-32 32-32l64 0c17.7 0 32 14.3 32 32l0 224 32 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 512c-17.7 0-32-14.3-32-32s14.3-32 32-32l32 0 0-192-32 0c-17.7 0-32-14.3-32-32z"
+                                />
+                            </svg>
+                        </button>
+                    </div>
                 </div>
             </div>
         </div>
@@ -919,7 +922,7 @@ export default {
 .legend-div {
     z-index: 2;
     width: 100%;
-    top: 70px;
+    top: 80px;
 }
 
 #info-button {
