+++ src/components/pages/StudentTidyTreeView.vue
@@ -1081,7 +1081,7 @@ export default {
 .legend-div {
     width: 100%;
     z-index: 2;
-    top: 70px;
+    top: 80px;
 }
 
 .bottom-legend-div {
