+++ src/components/pages/TidyTreeView.vue
@@ -1837,7 +1837,7 @@ export default {
 .legend-div {
     width: 100%;
     z-index: 2;
-    top: 70px;
+    top: 80px;
 }
 
 /* Modal filters */
